requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
torch>=2.0.0
transformers>=4.30.0
spacy>=3.6.0
PyMuPDF>=1.22.3
python-docx>=0.8.11
pydantic>=2.0.0
jinja2>=3.1.2